backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "faker>=37.3.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 120
